+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/PeripheralWorksCore.kt
@@ -9,6 +9,7 @@ import site.siredvin.peripheralworks.computercraft.*
 import site.siredvin.peripheralworks.computercraft.plugins.specific.SpecificPluginProvider
 import site.siredvin.peripheralworks.computercraft.plugins.specific.SpecificProtectedPluginProviders
 import site.siredvin.peripheralworks.data.ModText
+import site.siredvin.peripheralworks.networking.NetworkMessages
 import site.siredvin.peripheralworks.subsystem.entityperipheral.EntityPeripheralLookup
 import site.siredvin.peripheralworks.subsystem.entityperipheral.GenericEntityFluidStorageProvider
 import site.siredvin.peripheralworks.subsystem.entityperipheral.GenericEntityStorageProvider
@@ -18,6 +19,7 @@ import site.siredvin.peripheralworks.xplat.*
 @Suppress("UNUSED")
 object PeripheralWorksCore {
     const val MOD_ID = "peripheralworks"
+    const val NETWORK_VERSION = "1.0"
 
     var logger: Logger = LogManager.getLogger(MOD_ID)
 
@@ -33,7 +35,7 @@ object PeripheralWorksCore {
         ModPlatform.configure(platform)
         ModRecipeIngredients.configure(ingredients)
         ModBlocksReference.configure(blocks)
-        AgnosticItemStorageLookup.addItemStorageExtractor(MinecartUtils::minecartExtractor)
+        AgnosticItemStorageLookup.addBlockLookup(MinecartUtils::minecartExtractor)
         ComputerCraftProxy.addProvider(StorageProvider)
         ComputerCraftProxy.addProvider(FluidStorageProvider)
         ComputerCraftProxy.addProvider(EnergyStorageProvider)