+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/computercraft/GenericProvider.kt
@@ -23,7 +23,7 @@ object StorageProvider : PeripheralPluginProvider {
         if (!PeripheralWorksConfig.enableGenericItemStorage && !PeripheralWorksConfig.enableGenericInventory) {
             return null
         }
-        val storage = AgnosticItemStorageLookup.extractStorage(level, pos, level.getBlockEntity(pos)) ?: return null
+        val storage = AgnosticItemStorageLookup.extractFromBlock(level, pos, level.getBlockEntity(pos), side) ?: return null
         if (storage is SlottedAgnosticItemStorage && PeripheralWorksConfig.enableGenericInventory && storage.size != 0) {
             return InventoryPlugin(level, storage)
         }
@@ -40,7 +40,7 @@ object FluidStorageProvider : PeripheralPluginProvider {
 
     override fun provide(level: Level, pos: BlockPos, side: Direction): IPeripheralPlugin? {
         if (!PeripheralWorksConfig.enableGenericFluidStorage) return null
-        val storage = AgnosticFluidStorageLookup.extractFluidStorage(level, pos, level.getBlockEntity(pos)) ?: return null
+        val storage = AgnosticFluidStorageLookup.extractFromBlock(level, pos, level.getBlockEntity(pos), side) ?: return null
         return FluidStoragePlugin(level, storage, PeripheralWorksConfig.fluidStorageTransferLimit)
     }
 }
@@ -51,7 +51,7 @@ object EnergyStorageProvider : PeripheralPluginProvider {
 
     override fun provide(level: Level, pos: BlockPos, side: Direction): IPeripheralPlugin? {
         if (!PeripheralWorksConfig.enableGenericEnergyStorage) return null
-        val storage = AgnosticEnergyStorageLookup.extractEnergyStorage(level, pos, level.getBlockEntity(pos)) ?: return null
+        val storage = AgnosticEnergyStorageLookup.extractFromBlock(level, pos, level.getBlockEntity(pos), side) ?: return null
         if (PeripheralWorksConfig.energyAlwaysTransferable || EnergyRegistry.TRANSFERABLE.contains(storage.unit)) {
             return FullEnergyPlugin(level, storage, PeripheralWorksConfig.energyStorageTransferLimit)
         }