+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/computercraft/peripherals/RecipeRegistryPeripheral.kt
@@ -1,22 +1,28 @@
 package site.siredvin.peripheralworks.computercraft.peripherals
 
+import com.google.gson.JsonParser
 import dan200.computercraft.api.lua.IArguments
 import dan200.computercraft.api.lua.LuaException
 import dan200.computercraft.api.lua.LuaFunction
 import dan200.computercraft.api.lua.MethodResult
+import dan200.computercraft.shared.util.NBTUtil
+import net.minecraft.nbt.TagParser
 import net.minecraft.resources.ResourceLocation
 import net.minecraft.world.Container
 import net.minecraft.world.item.crafting.Recipe
 import net.minecraft.world.item.crafting.RecipeType
 import site.siredvin.broccolium.modules.platform.PlatformRegistries
+import site.siredvin.broccolium.modules.platform.PlatformToolkit
 import site.siredvin.peripheralworks.common.blockentity.RecipeRegistryBlockEntity
 import site.siredvin.peripheralworks.common.configuration.PeripheralWorksConfig
 import site.siredvin.peripheralworks.subsystem.recipe.RecipeRegistryToolkit
 import site.siredvin.tweakium.modules.peripheral.OwnedPeripheral
 import site.siredvin.tweakium.modules.peripheral.ext.getResourceLocation
 import site.siredvin.tweakium.modules.peripheral.owner.BlockEntityPeripheralOwner
-import java.util.*
+import java.io.BufferedReader
+import java.io.InputStreamReader
 import java.util.stream.Collectors
+import kotlin.jvm.optionals.getOrNull
 
 class RecipeRegistryPeripheral(
     blockEntity: RecipeRegistryBlockEntity,
@@ -31,9 +37,16 @@ class RecipeRegistryPeripheral(
     override val isEnabled: Boolean
         get() = PeripheralWorksConfig.enableRecipeRegistry
 
+    override val peripheralConfiguration: MutableMap<String, Any>
+        get() {
+            val base = super.peripheralConfiguration
+            base["apiVersion"] = listOf(2, 0)
+            return base
+        }
+
     @LuaFunction
     @Throws(LuaException::class)
-    fun getRecipeTypes(): MethodResult = MethodResult.of(
+    fun getTypes(): MethodResult = MethodResult.of(
         PlatformRegistries.RECIPE_TYPES.keySet().stream().filter {
             !RecipeRegistryToolkit.excludedRecipeTypes.contains(it)
         }.map(ResourceLocation::toString)
@@ -42,33 +55,49 @@ class RecipeRegistryPeripheral(
 
     @LuaFunction
     @Throws(LuaException::class)
-    fun getAllRecipesForType(arguments: IArguments): MethodResult {
-        val recipeTypeID: ResourceLocation = arguments.getResourceLocation(0)
-
-        @Suppress("UNCHECKED_CAST")
-        val type = PlatformRegistries.RECIPE_TYPES.tryGet(recipeTypeID) as? RecipeType<Recipe<Container>> ?: return MethodResult.of(false, "Cannot find recipe type $recipeTypeID")
-        return MethodResult.of(peripheralOwner.level!!.recipeManager.getAllRecipesFor(type).map { it.id.toString() })
+    fun get(recipeID: String): MethodResult {
+        val id = ResourceLocation.tryParse(recipeID) ?: return MethodResult.of(null)
+        val recipe = peripheralOwner.level!!.recipeManager.byKey(id).getOrNull() ?: return MethodResult.of(null)
+        return MethodResult.of(RecipeRegistryToolkit.serializeRecipe(recipe, peripheralOwner.level!!.registryAccess()))
     }
 
     @LuaFunction
-    @Throws(LuaException::class)
-    fun getRecipeForType(arguments: IArguments): MethodResult {
-        val recipeTypeID: ResourceLocation = arguments.getResourceLocation(0)
-        val recipeID: ResourceLocation = arguments.getResourceLocation(1)
+    fun getRaw(recipeID: String): MethodResult {
+        val recipeId = ResourceLocation.tryParse(recipeID) ?: return MethodResult.of(null)
 
-        @Suppress("UNCHECKED_CAST")
-        val type = PlatformRegistries.RECIPE_TYPES.tryGet(recipeTypeID) as? RecipeType<Recipe<Container>> ?: return MethodResult.of(false, "Cannot find recipe type $recipeTypeID")
-        return MethodResult.of(
-            peripheralOwner.level!!.recipeManager.getAllRecipesFor(type)
-                .filter { it.id == recipeID }
-                .map { RecipeRegistryToolkit.serializeRecipe(it, peripheralOwner.level!!.registryAccess()) }
-                .firstOrNull(),
+        @Suppress("DEPRECATION", "KotlinRedundantDiagnosticSuppress")
+        val recipePath = ResourceLocation(
+            recipeId.namespace,
+            "recipes/" + recipeId.path + ".json",
         )
+        val resourceManager = PlatformToolkit.get().minecraftServer?.resourceManager ?: return MethodResult.of(null, "Cannot find server variabel")
+        try {
+            val resource = resourceManager.getResource(recipePath).orElseThrow()
+            val reader = BufferedReader(
+                InputStreamReader(resource.open()),
+            )
+
+            val jsonObject = JsonParser.parseReader(reader).getAsJsonObject()
+            // Now you have the JSON content
+            reader.close()
+            val tag = TagParser.parseTag(jsonObject.toString())
+            return MethodResult.of(NBTUtil.toLua(tag))
+        } catch (e: Exception) {
+            return MethodResult.of(null, e.toString())
+        }
     }
 
     @LuaFunction
     @Throws(LuaException::class)
-    fun getRecipesFor(arguments: IArguments): MethodResult {
+    fun list(arguments: IArguments): MethodResult {
+        val types = arguments[0]
+        val recipeTypes = RecipeRegistryToolkit.collectRecipeTypes(types).toSet()
+        return MethodResult.of(peripheralOwner.level!!.recipeManager.recipes.filter { recipeTypes.contains(it.type) }.map { it.id.toString() })
+    }
+
+    @LuaFunction
+    @Throws(LuaException::class)
+    fun search(arguments: IArguments): MethodResult {
         val itemID: ResourceLocation = arguments.getResourceLocation(0)
         val types = arguments[1]
         val targetItem = PlatformRegistries.ITEMS.tryGet(itemID)