+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/networking/MessageType.kt
@@ -1,0 +1,4 @@
+package site.siredvin.peripheralworks.networking
+
+
+interface MessageType<T : NetworkMessage<*>>