+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/utils/MinecartUtils.kt
@@ -1,10 +1,12 @@
 package site.siredvin.peripheralworks.utils
 
 import net.minecraft.core.BlockPos
+import net.minecraft.core.Direction
 import net.minecraft.world.entity.vehicle.AbstractMinecart
 import net.minecraft.world.entity.vehicle.AbstractMinecartContainer
 import net.minecraft.world.level.Level
 import net.minecraft.world.level.block.Blocks
+import net.minecraft.world.level.block.entity.BlockEntity
 import net.minecraft.world.phys.AABB
 import site.siredvin.broccolium.modules.storage.item.ContainerWrapper
 import site.siredvin.broccolium.modules.storage.item.MergedContainer
@@ -22,19 +24,16 @@ object MinecartUtils {
         (pos.z + 1).toDouble() - SEARCH_MARGIN,
     )
 
-    fun getMinecarts(level: Level, pos: BlockPos): List<AbstractMinecart> = level.getEntitiesOfClass(AbstractMinecart::class.java, getSearchShape(pos))
+    fun getMinecarts(level: Level, pos: BlockPos): List<AbstractMinecart> = level.getEntitiesOfClass(AbstractMinecart::class.java, getSearchShape(pos)).sortedBy { it.uuid }
 
     fun getContainerMinecarts(level: Level, pos: BlockPos): List<AbstractMinecartContainer> = level.getEntitiesOfClass(AbstractMinecartContainer::class.java, getSearchShape(pos))
 
-    fun minecartExtractor(level: Level, obj: Any?): SlottedAgnosticItemStorage? {
-        if (obj !is BlockPos) {
-            return null
-        }
-        val state = level.getBlockState(obj)
+    fun minecartExtractor(level: Level, blockPos: BlockPos, blockEntity: BlockEntity?, direction: Direction?): SlottedAgnosticItemStorage? {
+        val state = level.getBlockState(blockPos)
         if (!state.`is`(Blocks.POWERED_RAIL)) {
             return null
         }
-        val containers = getContainerMinecarts(level, obj)
+        val containers = getContainerMinecarts(level, blockPos)
         if (containers.isEmpty()) {
             return null
         }