+++ projects/forge/src/main/kotlin/site/siredvin/peripheralworks/integrations/embers/Integration.kt
@@ -2,7 +2,6 @@ package site.siredvin.peripheralworks.integrations.embers
 
 import com.rekindled.embers.api.capabilities.EmbersCapabilities
 import site.siredvin.broccolium.modules.storage.energy.AgnosticEnergyStorageLookup
-import site.siredvin.broccolium.modules.storage.energy.api.AgnosticEnergyStorageExtractor
 import site.siredvin.peripheralworks.common.configuration.PeripheralWorksConfig
 
 class Integration : Runnable {
@@ -10,16 +9,19 @@ class Integration : Runnable {
     override fun run() {
         PeripheralWorksConfig.registerIntegrationConfiguration(Configuration)
         if (Configuration.enableEmberStorage) {
-            AgnosticEnergyStorageLookup.addEnergyStorageExtractor(
-                AgnosticEnergyStorageExtractor { level, blockPos, blockEntity ->
-                    if (blockEntity == null) return@AgnosticEnergyStorageExtractor null
+            AgnosticEnergyStorageLookup.addBlockLookup { level, blockPos, blockEntity, direction ->
+                    if (blockEntity == null) return@addBlockLookup null
                     val capability = blockEntity.getCapability(EmbersCapabilities.EMBER_CAPABILITY)
                     if (capability.isPresent) {
-                        return@AgnosticEnergyStorageExtractor AgnosticEmberStorage(capability.resolve().get())
+                        return@addBlockLookup AgnosticEmberStorage(capability.resolve().get())
                     }
-                    return@AgnosticEnergyStorageExtractor null
-                },
-            )
+                    if (direction != null) {
+                        val sidedCapability = blockEntity.getCapability(EmbersCapabilities.EMBER_CAPABILITY, direction)
+                        if (sidedCapability.isPresent)
+                            return@addBlockLookup AgnosticEmberStorage(sidedCapability.resolve().get())
+                    }
+                    return@addBlockLookup null
+                }
         }
     }
 }