+++ projects/forge/src/main/kotlin/site/siredvin/peripheralworks/integrations/emi/Entrypoint.kt
@@ -1,0 +1,13 @@
+package site.siredvin.peripheralworks.integrations.emi
+
+import dev.emi.emi.api.EmiEntrypoint
+import dev.emi.emi.api.EmiPlugin
+import dev.emi.emi.api.EmiRegistry
+
+@EmiEntrypoint
+class Entrypoint: EmiPlugin {
+    override fun register(p0: EmiRegistry) {
+        CommonEntrypoint.register(p0)
+    }
+
+}