+++ projects/forge/src/main/kotlin/site/siredvin/peripheralworks/integrations/mna/Integration.kt
@@ -4,7 +4,6 @@ import com.mna.entities.constructs.animated.Construct
 import net.minecraft.world.entity.Entity
 import site.siredvin.broccolium.modules.storage.item.AgnosticItemHandlerWrapper
 import site.siredvin.broccolium.modules.storage.item.AgnosticItemStorageLookup
-import site.siredvin.broccolium.modules.storage.item.api.AgnosticItemStorageEntityExtractor
 import site.siredvin.peripheralworks.subsystem.entityperipheral.EntityPeripheralLookup
 import site.siredvin.peripheralworks.subsystem.entityperipheral.EntityPeripheralPluginProvider
 import site.siredvin.tweakium.modules.peripheral.api.IPeripheralPlugin
@@ -23,13 +22,11 @@ class Integration : Runnable {
 
     override fun run() {
         EntityPeripheralLookup.addProvider(ConstructPluginProvider)
-        AgnosticItemStorageLookup.addItemStorageExtractor(
-            AgnosticItemStorageEntityExtractor { level, entity ->
-                if (entity is Construct) {
-                    return@AgnosticItemStorageEntityExtractor AgnosticItemHandlerWrapper(entity)
-                }
-                return@AgnosticItemStorageEntityExtractor null
-            },
-        )
+        AgnosticItemStorageLookup.addEntityLookup { level, entity, direction ->
+            if (entity is Construct) {
+                return@addEntityLookup AgnosticItemHandlerWrapper(entity)
+            }
+            return@addEntityLookup null
+        }
     }
 }