+++ projects/forge/src/main/kotlin/site/siredvin/peripheralworks/integrations/occultism/OccultismItemStorage.kt
@@ -1,5 +1,6 @@
 package site.siredvin.peripheralworks.integrations.occultism
 
+import com.klikli_dev.occultism.Occultism
 import com.klikli_dev.occultism.api.common.blockentity.IStorageController
 import net.minecraft.world.item.ItemStack
 import site.siredvin.broccolium.modules.storage.item.api.AgnosticItemStorage
@@ -7,6 +8,9 @@ import java.util.function.Predicate
 
 class OccultismItemStorage(private val storageController: IStorageController) : AgnosticItemStorage {
     override fun getItems(): Iterator<ItemStack> = storageController.stacks.iterator()
+
+    override val maxStackSize: Int
+        get() = if (Occultism.SERVER_CONFIG.storage.overrideItemStackSizes.get()) Occultism.SERVER_CONFIG.storage.controllerStackSize.get() else 64
 
     override fun setChanged() {
         storageController.onContentsChanged()